pyproject.toml
[project]
name = "pfs-fluxstd-lookup"
description = "Web application for matching astronomical catalogs with flux standard star catalogs from Gaia DR3 and Pan-STARRS DR2"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Masato Onodera", email = "monodera@naoj.org" },
  { name = "Subaru/PFS obsproc team", email = "pfs-obs-help@naoj.org" },
]
maintainers = [
  { name = "Masato Onodera", email = "monodera@naoj.org" },
  { name = "Subaru/PFS obsproc team", email = "pfs-obs-help@naoj.org" },
]
keywords = [
  "astronomy",
  "gaia",
  "pan-starrs",
  "pfs",
  "prime-focus-spectrograph",
  "spectroscopy",
  "web-application",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Astronomy",
]
dependencies = [
  "astropy-healpix>=1.1.2",
  "astropy[recommended]>=7.2.0",
  "duckdb>=1.4.3",
  "loguru>=0.7.3",
  "numpy>=2.3.5",
  "pandas[parquet]>=2.3.3",
  "panel>=1.8.4",
  "python-dotenv>=1.2.1",
]
dynamic = ["version"]

[dependency-groups]
dev = [
  "black>=25.12.0",
  "psycopg2-binary>=2.9.10",
  "ruff>=0.14.9",
  "ty>=0.0.3",
  "watchfiles>=1.1.1",
]

[build-system]
requires = ["hatch-vcs", "hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["app.py", "utils.py"]

[tool.hatch.version]
source = "vcs"
tag-pattern = "^v(?P<version>.+)$"
fallback-version = "0.0.0"

[tool.ruff]
target-version = "py312"
line-length = 88
force-exclude = true

[tool.ruff.lint]
# Explicitly include Ruff's default rule set (E4/E7/E9/F), then add extra rules.
select = ["E4", "E7", "E9", "F", "B", "I", "UP", "D"]
ignore = [
  "E501",  # line too long (handled by Black)
  "F401",  # imported but unused (too noisy with many imports)
  "D103",  # Missing docstring in public function
  "D100",  # Missing docstring in public module
]
# Allow Ruff to apply all available auto-fixes when running with `--fix`.
fixable = ["ALL"]

[tool.ruff.lint.pydocstyle]
# Enforce NumPy-style docstrings.
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
# Keep tests lightweight: don't require docstrings in test files.
"tests/**.py" = ["D"]
